% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{BuildEstimator}
\alias{BuildEstimator}
\title{Build estimator given sensitivity}
\usage{
BuildEstimator(k, eo, B, M, K, p = Inf, alpha = 0.05)
}
\arguments{
\item{k}{Sensitivity, or matrix of sensitivities}

\item{p}{1, 2, or Inf}
}
\description{
Build estimator given sensitivity
}
\keyword{internal}
