% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{EffBounds}
\alias{EffBounds}
\title{Efficiency bounds under \eqn{\ell_p}{ell_p} constraints}
\usage{
EffBounds(eo, B, M, p = 2, beta = 0.5, alpha = 0.05)
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with
       dimension \eqn{d_g} by \eqn{d_g}, where \eqn{d_g} is the number of
       moments}

    \item{G}{Estimate of derivative of the moment condition, matrix with
    dimension \eqn{d_g} by \eqn{d_\theta}{d_theta}, where
    \eqn{d_\theta}{d_theta} is the dimension of \eqn{\theta}{theta}}

    \item{H}{Estimate of derivative of \eqn{h(\theta)}{h(theta)}. A vector of
    length \eqn{d_\theta}{d_theta}}

    \item{n}{sample size}

    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension \eqn{d_g} by
\eqn{d_\gamma}{d_gamma} that determines the set \eqn{\mathcal{C}}, where
\eqn{d_\gamma}{d_gamma} is the number of invalid moments, and \eqn{d_g}
is the number of moments}

\item{M}{Bound on the norm of \eqn{\gamma}{gamma}}

\item{p}{Parameter determining which \eqn{\ell_p}{lp} norm to use, must equal
\code{1}, \code{2}, or \code{Inf}.}

\item{beta}{Quantile of excess length that a one-sided confidence interval is
optimizing.}

\item{alpha}{determines confidence level, \code{1-alpha}, for
constructing/optimizing confidence intervals.}
}
\value{
A list with two elements, \code{"onesided"} for efficiency of
    one-sided CIs and \code{"twosided"} for efficiency of two-sided CIs
}
\description{
Computes the asymptotic efficiency of two-sided fixed-length confidence
intervals at \eqn{c=0}, as well as the efficiency of one-sided confidence
intervals that optimize a given \code{beta} quantile of excess length, when
the set \eqn{\mathcal{C}} is characterized by \eqn{\ell_p} constraints.
}
\details{
The set \eqn{\mathcal{C}} takes the form \eqn{B\gamma}{B*gamma} where the
\eqn{\ell_p}{lp} norm of \eqn{\gamma}{gamma} is bounded by \eqn{M}.
}
\examples{
## Replicates first line of Table 2 in Armstrong and Kolesár (2020)
## First compute matrix B
I <- vector(mode="logical", length=nrow(blp$G))
I[6] <- TRUE
B <- (blp$ZZ \%*\% diag(sqrt(blp$n)*abs(blp$perturb)/blp$sdZ))[, I, drop=FALSE]
eo <- list(H=blp$H, G=blp$G, Sig=solve(blp$W), n=blp$n, g_init=blp$g_init)
EffBounds(eo, B, M=1, p=Inf, beta=0.5, alpha=0.05)
}
\references{
{

\cite{Armstrong, T. B., and M. Kolesár (2020): Sensitivity Analysis Using
Approximate Moment Condition Models,
\url{https://arxiv.org/abs/1808.07387}}

}
}
