% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.R
\name{Jtest}
\alias{Jtest}
\title{J-test of overidentifying restrictions under local misspecification}
\usage{
Jtest(eo, B, M = 1, p = 2, alpha = 0.05)
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with
       dimension \eqn{d_g} by \eqn{d_g}, where \eqn{d_g} is the number of
       moments}

    \item{G}{Estimate of derivative of the moment condition, matrix with
    dimension \eqn{d_g} by \eqn{d_\theta}{d_theta}, where
    \eqn{d_\theta}{d_theta} is the dimension of \eqn{\theta}{theta}}

    \item{n}{sample size}

    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension \eqn{d_g} by
\eqn{d_\gamma}{d_gamma} that determines the set \eqn{\mathcal{C}}, where
\eqn{d_\gamma}{d_gamma} is the number of invalid moments, and \eqn{d_g}
is the number of moments}

\item{M}{Bound on the norm of \eqn{\gamma}{gamma}}

\item{p}{Parameter determining which \eqn{\ell_p}{lp} norm to use, must equal
\code{1}, \code{2}, or \code{Inf}.}

\item{alpha}{determines confidence level, \code{1-alpha}, for
constructing/optimizing confidence intervals.}
}
\value{
List with three elements: \describe{

\item{J}{Value of J statistic}

\item{p0}{P-value of usual J test}

\item{pC}{P-value for J-test that  allows for local misspecification}

\item{Mmin}{Minimum value of \eqn{M} for which the J-test does not reject} }
}
\description{
Computes J-test of overidentifying restrictions with critical value adjusted
to allow for local misspecification, when the parameter \eqn{c} takes the
form \eqn{c=B\gamma}{c=B*gamma} with the \eqn{\ell_p}{lp} norm of
\eqn{\gamma}{gamma} bounded by \eqn{M}.
}
\details{
The test assumes initial estimator in \code{eo} is optimal under correct
specification, computed using \code{eo$Sig} as the weight matrix. The test is
based on a J statistic using critical values that account for local
misspecification; see appendix B in Armstrong and Kolesár (2020) for details.
}
\examples{
## Replicates first line of Table 1 in Armstrong and Kolesár (2020)
## 1. Compute matrix B when instrument D/F # cars is invalid
I <- vector(mode="logical", length=nrow(blp$G))
I[6] <- TRUE
B <- (blp$ZZ \%*\% diag(sqrt(blp$n)*abs(blp$perturb)/blp$sdZ))[, I, drop=FALSE]
## 2. Make sure Sig corresponds to inverse of weight matrix
eo <- list(G=blp$G, Sig=solve(blp$W), n=blp$n, g_init=blp$g_init)
Jtest(eo, B, M=1, p=2, alpha=0.05)
Jtest(eo, B, M=1, p=Inf, alpha=0.05)
}
\references{
{

\cite{Armstrong, T. B., and M. Kolesár (2020): Sensitivity Analysis Using
Approximate Moment Condition Models,
\url{https://arxiv.org/abs/1808.07387}}

}
}
