% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm1.R
\name{l1h0}
\alias{l1h0}
\title{Orthogonalized homotopy solution for l_1}
\usage{
l1h0(G, Sig, h, I)
}
\arguments{
\item{I}{vector of indicators which instruments are invalid}
}
\description{
Orthogonalized homotopy solution for l_1
}
\keyword{internal}
