% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm2.R
\name{l2sens}
\alias{l2sens}
\title{Construct optimal sensitivity and compute worst-case bias and sd for a given
lambda}
\usage{
l2sens(eo, B, M = diag(ncol(B)), lam, K, alpha = 0.05)
}
\arguments{
\item{eo}{object containig parameter estimates based on theta_init}

\item{B}{matrix}

\item{M}{matrix}

\item{K}{diameter}

\item{alpha}{for CI coverage}

\item{lambda}{weight on bias}
}
\description{
Construct optimal sensitivity and compute worst-case bias and sd for a given
lambda
}
\keyword{internal}
