% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmminf.R
\name{linfh0}
\alias{linfh0}
\title{Orthogonalized homotopy solution for l_infty}
\usage{
linfh0(G, Sig, H, I)
}
\arguments{
\item{I}{vector of indicators which instruments are invalid}
}
\description{
Orthogonalized homotopy solution for l_infty
}
\keyword{internal}
