% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{modulus}
\alias{modulus}
\title{Modulus}
\usage{
modulus(delta, eo, B, M, p = 2, spath = NULL)
}
\description{
Modulus
}
\keyword{Bk=0,}
\keyword{If}
\keyword{c_delta}
\keyword{c_delta.}
\keyword{delta.}
\keyword{give}
\keyword{internal}
\keyword{largest}
\keyword{leads}
\keyword{possible}
\keyword{that}
\keyword{to}
