% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{OptEstimator}
\alias{OptEstimator}
\title{One-step estimator based on optimal sensitivity under ell_p constraints}
\usage{
OptEstimator(eo, B, M = diag(ncol(B)), K, p = 2, res = NULL,
  alpha = 0.05, opt.criterion = "FLCI")
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with dimension
       d_g by d_g, where d_g is the number of moments}

    \item{G}{Estimate of defivative of the moment condition, matrix with
    dimension d_g by d_theta, where d_theta is the dimension of \eqn{theta}}

    \item{H}{Estimate of defivative of \eqn{h(theta_{0})}. A vector of length d_theta}
    \item{n}{sample size}
    \item{h_init}{Estimate of \eqn{h(theta_0)}}
    \item{k_init}{Initial sensitivity}
    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension d_g by d_gamma that
determines the set \eqn{C}, where d_gamma is the number of invalid
moments, and d_g is the number of moments}

\item{M}{invertible matrix \eqn{M} with dimensions d_gamma by d_gamma that
determines the set \eqn{C}.}

\item{K}{diameter of set \eqn{C}}

\item{p}{Parameter determining which ell_p norm to use, one of \code{1},
\code{2}, or \code{Inf}.}

\item{res}{Optionally, the solution path, output of \code{lph} to speed up
computation. For \code{p==1} and \code{p==Inf} only.}

\item{alpha}{determines confidence level, \code{1-alpha}, for
constructing/optimizing confidence intervals.}

\item{opt.criterion}{Optimality criterion for choosing optimal bias-variance
    tradeoff. The options are:

   \describe{

   \item{\code{"MSE"}}{Mean squared errror}

   \item{\code{"FLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"Valid"}}{Optimal estimator under valid moments. This returns
   the original estimator, with confidence intervals adjusted for possible
   misspecification}

    }}
}
\description{
Computes the optimal sensitivity and the optimal estimator when the set
\eqn{C} takes the form \eqn{B*gamma} where the ell_p norm of M*gamma is
bounded by K.
}
