% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{blp}
\alias{blp}
\title{Estimates from Berry, Levinsohn, and Pakes (1995)}
\format{
A list, consisting 11 objects:

\describe{

\item{G}{Matrix with 31 rows and 17 columns, estimate of derivative of the
         moment condition evaluated at initial estimate of \eqn{\theta} from
         Berry, Levinsohn, and Pakes (1995),
         \eqn{\hat{\theta}_{initial}}{thetahat_init}.}

\item{H}{Vector of length 17, estimate of derivative of average markup
  \eqn{h(\theta)}{h(theta)} with respect to model parameters
  \eqn{\theta}{theta}, evaluated at
  \eqn{\hat{\theta}_{initial}}{thetahat_init}.}

\item{W}{Weight matrix used to obtain the estimate
         \eqn{\hat{\theta}_{initial}}{thetahat_init}, preliminary estimate of
         variance of moment conditions.}

\item{g_init}{Average moment condition, evaluated at
              \eqn{\hat{\theta}_{initial}}{thetahat_init}.}

\item{h_init}{Estimate of the average markup,
              \eqn{h(\hat{\theta}_{initial})}{h(thetahat_init)}.}

\item{names}{Two lists, one for names of the moment conditions, and one for
             elements of \eqn{\theta}{theta}.}

\item{ZZ}{Gram matrix \eqn{Z'Z} of the instruments, used to specify the
misspecification set \eqn{\mathcal{C}}.}

\item{Sig}{Estimate of variance of moment condition.}

\item{sdZ}{Vector of standard deviations of the instruments.}

\item{perturb}{scaling parameters to give interpretable meaning to violations
of supply-side conditions. See vignette \code{vignette("GMMSensitivity")} for
details.}

\item{n}{Sample size, number of car models.}

}

See Armstrong and Kolesár (2020) for a detailed description of these objects.
}
\source{
Replication files for Andrews, Gentzkow, and Shapiro (2017),
    available at \url{https://doi.org/10.7910/DVN/LLARSN}
}
\usage{
blp
}
\description{
This dataset contains estimates of the model in Berry, Levinsohn, and Pakes
(1995), as implemented by Andrews, Gentzkow, and Shapiro (2017). It is used
to illustrate the confidence intervals implemented in this package.
}
\references{
{

\cite{Andrews, I., M. Gentzkow, and J. M. Shapiro (2017): Measuring the
Sensitivity of Parameter Estimates to Sample Statistics, Quarterly Journal of
Economics, 132, 1553–1592.}

\cite{Armstrong, T. B., and M. Kolesár (2020): Sensitivity Analysis Using
Approximate Moment Condition Models, \url{https://arxiv.org/abs/1808.07387}}

\cite{Berry, S. T., J. Levinsohn, and A. Pakes (1995): Automobile Prices in
Market Equilibrium, Econometrica, 63, 841–890.}

}
}
\keyword{datasets}
