% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{blp}
\alias{blp}
\title{Estimates from Berry, Levinsohn, and Pakes (1995)}
\format{A list, consisting 11 objects:

\describe{

\item{G}{Matrix with 31 rows and 17 columns, estimate of derivative of the moment condition.}

\item{H}{Vector of length 17, estimate of derivative of average markup
  \eqn{h(theta)} with respect to model parameters theta.}

\item{W}{Weight matrix used to obtain the estimate of theta.}

\item{g_init}{Average moment condition, evaluated at estimate \eqn{theta_init} of theta}

\item{h_init}{Estimate of the average markup, \eqn{h(theta_init)}}

\item{names}{Two lists, one for names of the moment conditions, and one for elements of theta}

\item{OmZZ}{Gram matrix of the instruments, used to specify the set script C.}

\item{Sig}{Variance of moment condition}

\item{sdZ}{vector of standard deviations of the instruments}

\item{perturb}{vector of perturbations \eqn{delta} used to specify the set script C.}

\item{n}{Sample size, number of model/years.}
}

See Armstrong and Kolesár (2018) for a detailed description of these objects.}
\source{
Replication files for Andrews, Gentzkow, and Shapiro (2017), available at
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/LLARSN/2KFPRA&version=1.1}
}
\usage{
blp
}
\description{
This dataset contains estimates of the model in Berry, Levinsohn, and Pakes
(1995), as implemented by Andrews, Gentzkow, and Shapiro (2017), and it is used
to illustrate the confidence intervals implemented in this package.
}
\references{
{

\cite{Andrews, I., M. Gentzkow, and J. M. Shapiro (2017): "Measuring the
Sensitivity of Parameter Estimates to Sample Statistics," Quarterly Journal
of Economics, 132, 1553–1592.}

\cite{Armstrong, T. B., and M. Kolesár (2018):
"Sensitivity Analysis Using Approximate Moment Condition Models," Unpublished manuscript}

\cite{Berry, S. T., J. Levinsohn, and A. Pakes (1995): "Automobile Prices in
Market Equilibrium," Econometrica, 63, 841–890.}


}
}
\keyword{datasets}
