% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{EffBounds}
\alias{EffBounds}
\title{Efficiency bounds under \eqn{\ell_p}{ell_p} constraints}
\usage{
EffBounds(eo, B, M, p = 2, beta = 0.5, alpha = 0.05, cvx = FALSE)
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with dimension
       \eqn{d_g} by \eqn{d_g}, where \eqn{d_g} is the number of moments}

    \item{G}{Estimate of derivative of the moment condition, matrix with
    dimension \eqn{d_g} by \eqn{d_\theta}{d_theta}, where
    \eqn{d_\theta}{d_theta} is the dimension of \eqn{\theta}{theta}}

    \item{H}{Estimate of derivative of \eqn{h(\theta)}{h(theta)}. A vector of
    length \eqn{d_\theta}{d_theta}}

    \item{n}{sample size}

    \item{h_init}{Estimate of \eqn{h(\theta)}{h(theta)}}

    \item{k_init}{Initial sensitivity}

    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension \eqn{d_g} by
\eqn{d_\gamma}{d_gamma} that determines the set \eqn{C}, where
\eqn{d_\gamma}{d_gamma} is the number of invalid moments, and \eqn{d_g}
is the number of moments}

\item{M}{Bound on the norm of \eqn{\gamma}{gamma}}

\item{p}{Parameter determining which \eqn{\ell_p}{lp} norm to use, must equal
\code{1}, \code{2}, or \code{Inf}.}

\item{beta}{Quantile of excess length that a one-sided confidence interval is
optimizing.}

\item{alpha}{determines confidence level, \code{1-alpha}, for
constructing/optimizing confidence intervals.}

\item{cvx}{By default, the efficiency for \code{p=1} and for \code{p=Inf} is
computed using the homotopy algorithm described in Appendix A of
Armstrong and Kolesár (2018). If \code{cvx=TRUE} is specified, the
modulus is computed using the CVX convex optimizer. This option is
included mostly just to verify that the homotopy solution correct.}
}
\description{
Computes the asymptotic efficiency of two-sided fixed-length confidence
intervals at \eqn{c=0}, as well as the efficiency of one-sided confidence
intervals that optimize a given \code{beta} quantile of excess length, when
the set \eqn{C} is characterized by \eqn{\ell_p} constraints.
}
\details{
The set \eqn{C} takes the form \eqn{B\gamma}{B*gamma} where the
\eqn{\ell_p}{lp} norm of \eqn{\gamma}{gamma} is bounded by \eqn{M}.
}
\references{
{

\cite{Armstrong, T. B., and M. Kolesár (2018): Sensitivity Analysis
    Using Approximate Moment Condition Models, Unpublished manuscript}
}
}
