% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.R
\name{Jtest}
\alias{Jtest}
\title{J-test of overidentifying restrictions under local misspecification}
\usage{
Jtest(eo, B, K, p = 2, alpha = 0.05)
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with dimension
       d_g by d_g, where d_g is the number of moments}

    \item{G}{Estimate of defivative of the moment condition, matrix with
    dimension d_g by d_theta, where d_theta is the dimension of \eqn{theta}}

    \item{H}{Estimate of defivative of \eqn{h(theta_{0})}. A vector of length d_theta}
    \item{n}{sample size}
    \item{h_init}{Estimate of \eqn{h(theta_0)}}
    \item{k_init}{Initial sensitivity}
    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension d_g by d_gamma that
determines the set \eqn{C}, where d_gamma is the number of invalid
moments, and d_g is the number of moments}

\item{K}{diameter of set \eqn{C}}

\item{p}{Parameter determining which ell_p norm to use, one of \code{1},
\code{2}, or \code{Inf}.}

\item{alpha}{determines confidence level, \code{1-alpha}, for
constructing/optimizing confidence intervals.}
}
\value{
List with three elements: \describe{

\item{p0}{Value of J statistic}

\item{p0}{P-value of usual J test}

\item{pC}{P-value for J-test that  allows for local misspecification}

\item{Kmin}{Minimum value of \code{K} for which the J-test would not reject} }
}
\description{
Computes J-test of overidentifying restrictions with critical value adjusted
to allow for local misspecification, when the parameter \eqn{c} takes the
form \eqn{c=B*gamma} with the ell_p norm of gamma is bounded by K. Assumes
initial estimator in \code{eo} is optimal under correct specification.
}
