% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm2.R
\name{l2opt}
\alias{l2opt}
\title{One-step estimator based on optimal sensitivity under ell_2 constraints}
\usage{
l2opt(eo, B, M, alpha = 0.05, opt.criterion = "FLCI")
}
\description{
One-step estimator based on optimal sensitivity under ell_2 constraints
}
\keyword{internal}
