% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm2.R
\name{l2sens}
\alias{l2sens}
\title{Construct optimal sensitivity and compute worst-case bias and sd for a given
kappa}
\usage{
l2sens(eo, B, kap, M, alpha = 0.05)
}
\description{
Construct optimal sensitivity and compute worst-case bias and sd for a given
kappa
}
\keyword{internal}
