% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmminf.R
\name{lph}
\alias{lph}
\title{Compute solution path for l_infty or l_1 constraints}
\usage{
lph(eo, B, M = diag(ncol(B)), p = Inf)
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with dimension
       d_g by d_g, where d_g is the number of moments}

    \item{G}{Estimate of defivative of the moment condition, matrix with
    dimension d_g by d_theta, where d_theta is the dimension of \eqn{theta}}

    \item{H}{Estimate of defivative of \eqn{h(theta_{0})}. A vector of length d_theta}
    \item{n}{sample size}
    \item{h_init}{Estimate of \eqn{h(theta_0)}}
    \item{k_init}{Initial sensitivity}
    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension d_g by d_gamma that
determines the set \eqn{C}, where d_gamma is the number of invalid
moments, and d_g is the number of moments}

\item{M}{invertible matrix \eqn{M} with dimensions d_gamma by d_gamma that
determines the set \eqn{C}.}

\item{p}{Parameter determining which ell_p norm to use, one of \code{1}, or
\code{Inf}.}
}
\description{
Computes the vector of optimal sensitivities at each knot of the solution
path that traces out the optimal bias-variance frontier when the set \eqn{C}
takes the form \eqn{B*gamma}, with the ell_p norm of M*gamma bounded by a
constant, for \eqn{p=1}, or \eqn{p=Inf}. This path is used as an input to
\code{\link{OptEstimator}}.
}
