% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{modulus}
\alias{modulus}
\title{c_delta. If Bk=0, give c_delta that leads to largest possible delta.}
\usage{
modulus(delta, eo, B, M, p = 2, spath = NULL)
}
\description{
c_delta. If Bk=0, give c_delta that leads to largest possible delta.
}
