% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{OptEstimator}
\alias{OptEstimator}
\title{One-step estimator based on optimal sensitivity under \eqn{\ell_p}{lp}
constraints}
\usage{
OptEstimator(
  eo,
  B,
  M,
  p = 2,
  spath = NULL,
  alpha = 0.05,
  opt.criterion = "FLCI"
)
}
\arguments{
\item{eo}{List containing initial estimates with the following components:

    \describe{

    \item{Sig}{Estimate of variance of the moment condition, matrix with
       dimension \eqn{d_g} by \eqn{d_g}, where \eqn{d_g} is the number of
       moments}

    \item{G}{Estimate of derivative of the moment condition, matrix with
    dimension \eqn{d_g} by \eqn{d_\theta}{d_theta}, where
    \eqn{d_\theta}{d_theta} is the dimension of \eqn{\theta}{theta}}

    \item{H}{Estimate of derivative of \eqn{h(\theta)}{h(theta)}. A vector of
    length \eqn{d_\theta}{d_theta}}

    \item{n}{sample size}

    \item{h_init}{Estimate of \eqn{h(\theta)}{h(theta)}}

    \item{k_init}{Initial sensitivity}

    \item{g_init}{Moment condition evaluated at initial estimate}

    }}

\item{B}{matrix \eqn{B} with full rank and dimension \eqn{d_g} by
\eqn{d_\gamma}{d_gamma} that determines the set \eqn{C}, where
\eqn{d_\gamma}{d_gamma} is the number of invalid moments, and \eqn{d_g}
is the number of moments}

\item{M}{Bound on the norm of \eqn{\gamma}{gamma}}

\item{p}{Parameter determining which \eqn{\ell_p}{lp} norm to use, must equal
\code{1}, \code{2}, or \code{Inf}.}

\item{spath}{Optionally, the solution path, output of \code{lph} to speed up
computation. For \code{p==1} and \code{p==Inf} only.}

\item{alpha}{determines confidence level, \code{1-alpha}, for
constructing/optimizing confidence intervals.}

\item{opt.criterion}{Optimality criterion for choosing optimal bias-variance
    tradeoff. The options are:

   \describe{

   \item{\code{"MSE"}}{Minimize worst-case mean squared error of the
   estimator.}

   \item{\code{"FLCI"}}{Length of (fixed-length) two-sided confidence
       intervals.}

   \item{\code{"Valid"}}{Optimal estimator under valid moments. This returns
   the original estimator, with confidence intervals adjusted for possible
   misspecification}

    }}
}
\description{
Computes the optimal sensitivity and the optimal estimator when the set
\eqn{C} takes the form \eqn{c=B\gamma}{c=B*gamma} with the \eqn{\ell_p}{lp}
norm of \eqn{\gamma}{gamma} bounded by \eqn{M}.
}
